+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 * Added: New task ZipFile for local deployment to a ZIP file in the bin folder
+* Changed: The temporary debug script should be used by the VS Code debug task
 
 ## 2.3.0 - 2020-06-02